+++ src/CHANGELOG.md
@@ -19,6 +19,7 @@
 - Utils: Add spaCy’s Amharic, Kyrgyz, Macedonian, Sanskrit, Tigrinya, and Tswana word tokenizers
 - Utils: Add spaCy’s Amharic, Korean, Kyrgyz, Macedonian, Sanskrit, Tigrinya, and Tswana stop word lists
 - Utils: Add spaCy’s Russian POS tagger and lemmatizer
+- Utils: Add spaCy’s sentence recognizer
 - Work Area: Add Concordancer (Parallel Mode)
 
 ### 📌 Bugfixes